Add files via upload
begun merging the 3d vector class as a derivitive of my 2d vector class. I am going to take the overloaded operators that are not yet in the 2d class over to the base class so I can make them work there as well and then they will be redefined in the defived class to work with  x,y,z not jsut x,y.